moved the fbo down one level
