+++ common/src/main/kotlin/com/lambda/graphics/buffer/FrameBuffer.kt
@@ -1,4 +1,5 @@
-package com.lambda.graphics.buffer.fbo
+package com.lambda.graphics.buffer
+
 import com.lambda.Lambda.mc
 import com.lambda.graphics.RenderMain
 import com.lambda.graphics.buffer.vao.VAO
